+++ theory/CudaGeneralizedLorenzMie.py
@@ -3,8 +3,8 @@
 
 import numpy as np
 import cupy as cp
-from pylorenzmie.theory.cukernels import cufield, cufieldf
-from pylorenzmie.theory.fastkernels import fastfield
+from pylorenzmie.theory.cuholo import cufield, cufieldf
+from pylorenzmie.theory.fastholo import fastfield
 from pylorenzmie.theory.GeneralizedLorenzMie import GeneralizedLorenzMie
 
 cp.cuda.Device()
